Fix SLURM /etc/sysconfig/slurm settings script
The SLURM startup config file/script in /etc/sysconfig/slurm is sourced
when SLURM daemons are starting and stopping, so we need to check for
the current state before checking the gres.conf file. We don't want the
script to be checking for GPUs/Phis during a SLURM shutdown.

The code which was setting CPU affinity for GPUs had a typo.
Additionally, SchedMD recommends that such affinity settings be left off
unless they are important. Checking CPU/GPU affinity slows down
scheduling, so we shouldn't use that feature by default.